+++ tezt/lib_tezos/proxy_server.ml
@@ -27,6 +27,7 @@ type argument =
   (* If you're considering adding [--endpoint] and [--rpc-addr],
      beware that, for the moment, they are automatically computed in
      {!create} below *)
+  | Data_dir
   | Symbolic_block_caching_time of int
 
 module Parameters = struct
@@ -69,6 +70,7 @@ let create ?runner ?name ?rpc_port ?(args = []) node =
   let user_arguments =
     List.map
       (function
+        | Data_dir -> ["--data-dir"; Node.data_dir node]
         | Symbolic_block_caching_time s ->
             ["--sym-block-caching-time"; Int.to_string s])
       args